stream/requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.6
django-rest-swagger==2.1.2
djangorestframework==3.7.1
docker==3.5.1
docker-pycreds==0.3.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
websocket-client==0.53.0
